requirements.txt
flask
pymongo
PyGithub